+++ frontend/src/hooks/useCreateCustomerLocally.ts
@@ -39,6 +39,7 @@ const useCreateCustomerLocally = () => {
                     // customer will handle this section
                 }
             );
+            console.log(data);
             if (data)
                 toast.success(`${data?.message}`, {
                     duration: 8000,